Code clean up

+++ keramics-formats/src/fat/file_entry.rs
@@ -131,7 +131,7 @@ impl FatFileEntry {
                 );
                 match block_reader.open(
                     &self.block_allocation_table,
-                    directory_entry.short_name.data_start_cluster as u32,
+                    directory_entry.short_name.data_start_cluster,
                 ) {
                     Ok(_) => {}
                     Err(mut error) => {
@@ -210,7 +210,7 @@ impl FatFileEntry {
     /// Reads the sub directory entries.
     fn read_sub_directory_entries(&mut self) -> Result<(), ErrorTrace> {
         let cluster_block_number: u32 = match &self.directory_entry {
-            Some(directory_entry) => directory_entry.short_name.data_start_cluster as u32,
+            Some(directory_entry) => directory_entry.short_name.data_start_cluster,
             None => {
                 return Err(keramics_core::error_trace_new!("Missing directory entry"));
             }
